Fix CI test, minor cleanup.
