+++ HPXMLtoOpenStudio/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.1</schema_version>
   <name>hpxm_lto_openstudio</name>
   <uid>b1543b30-9465-45ff-ba04-1d1f85e763bc</uid>
-  <version_id>3d46ea32-e18d-4f8d-b56d-e8387dab96e9</version_id>
-  <version_modified>2025-03-28T16:59:26Z</version_modified>
+  <version_id>c48d5851-7cbd-4c8a-81d5-bbed955b49d6</version_id>
+  <version_modified>2025-03-28T22:25:17Z</version_modified>
   <xml_checksum>D8922A73</xml_checksum>
   <class_name>HPXMLtoOpenStudio</class_name>
   <display_name>HPXML to OpenStudio Translator</display_name>
@@ -213,7 +213,7 @@
       <filename>constructions.rb</filename>
       <filetype>rb</filetype>
       <usage_type>resource</usage_type>
-      <checksum>75FBCCF4</checksum>
+      <checksum>7745F4F8</checksum>
     </file>
     <file>
       <filename>data/Xing_okstate_0664D_13659_Table_A-3.csv</filename>
@@ -351,7 +351,7 @@
       <filename>geometry.rb</filename>
       <filetype>rb</filetype>
       <usage_type>resource</usage_type>
-      <checksum>9FE4D0B1</checksum>
+      <checksum>B9F6B522</checksum>
     </file>
     <file>
       <filename>hotwater_appliances.rb</filename>
@@ -423,7 +423,7 @@
       <filename>materials.rb</filename>
       <filetype>rb</filetype>
       <usage_type>resource</usage_type>
-      <checksum>23B30F2C</checksum>
+      <checksum>0881326D</checksum>
     </file>
     <file>
       <filename>math.rb</filename>
@@ -717,7 +717,7 @@
       <filename>test_enclosure.rb</filename>
       <filetype>rb</filetype>
       <usage_type>test</usage_type>
-      <checksum>9D653777</checksum>
+      <checksum>EA90A86D</checksum>
     </file>
     <file>
       <filename>test_generator.rb</filename>
